refactor: estilos das páginas de autenticação movido para arquivo compartilhado authstyles
