+++ src/styles/shared/authStyles.ts
@@ -1,0 +1,2 @@
+export const mainContainer = "bg-carbone-secondary h-screen w-full p-16 flex-row"
+export const content = "flex h-full bg-carbone justify-around p-16 items-center rounded-lg"
